Move pending vote management from VertexStore to designated PendingVotes
